requirements.txt
fastapi
uvicorn[standard]
google-generativeai
sentence-transformers
scikit-learn
numpy
matplotlib
chromadb
pydantic
python-multipart
